Modified the ERD and the schema text to match the database tables.
Changed the ERD.svg to .pdf to see if it will fix the display problem.

Testing the CSS stylesheet.

Added formatting to the ERD, and continue to have issues with the ERD.svg image link within the document.

Removed the CSS stylesheet on the personas.php document.

Dylan and Rochelle waved their magic wands to correct the display. Used browse remote host from PHPStorm.